requirements.txt
carvekit==4.1.2
fastapi==0.115.6
pillow==11.1.0
pydantic==2.10.5
uvicorn==0.34.0
pydantic_settings==2.7.1
cloudinary==2.3.0